Cast captcha randomise() loop bounds to int
PHP 8 floor()/ceil() return floats, so $i was a float when used as a
string offset in $scanline[$i], emitting a "String offset cast occurred"
warning. That output then broke the subsequent header() call with
"headers already sent" during registration.

+++ forums/includes/usercp_confirm.php
@@ -155,8 +155,8 @@ exit;
 function randomise($scanline, $width)
 {
 	$new_line = '';
-	$start = floor($width/2);
-	$end = strlen($scanline) - ceil($width/2);
+	$start = (int) floor($width/2);
+	$end = strlen($scanline) - (int) ceil($width/2);
 
 	for ($i = $start; $i < $end; $i++)
 	{
